+++ elite_dangerous_colonise/Models/Database Types/UncolonisedDetailsInsertType.cs
@@ -6,7 +6,7 @@ namespace elite_dangerous_colonise.Models.Database_Types
     public class UncolonisedDetailsInsertType
     {
         [PgName("systemID")]
-        public long SystemID { get; set; }
+        public decimal SystemID { get; set; }
         [PgName("lastUpdated")]
         public DateTime LastUpdate { get; set; }
         [PgName("reserveLevel")]
@@ -54,7 +54,7 @@ namespace elite_dangerous_colonise.Models.Database_Types
 
         public UncolonisedDetailsInsertType() { }
         public UncolonisedDetailsInsertType(
-            long systemID, DateTime lastUpdate, ReserveType reserveLevel, short landableCount, short walkableCount,
+            ulong systemID, DateTime lastUpdate, ReserveType reserveLevel, short landableCount, short walkableCount,
             short totalHotspots, double systemValue, short blackHoleCount, short neutronStarCount, short whiteDwarves, short otherStarCount,
             short earthLikeCount, short waterWorldCount, short ammoniaWorldCount, short gasGiantCount, short highMetalContentCount, short metalRichCount,
             short rockyIceBodyCount, short rockBodyCount, short icyBodyCount, short organicCount, short geologicalsCount, short ringCount
